Move fold range conversion to ext
