+++ crates/djls-ide/src/folding.rs
@@ -1,39 +1,20 @@
 use djls_semantic::structure::forest::SemanticNode;
 use djls_source::File;
-use djls_source::LineIndex;
 use djls_source::Span;
 use djls_templates::Node;
 use tower_lsp_server::ls_types;
 
-use crate::ext::FoldingRangeKindExt;
-use crate::ext::SpanExt;
+use crate::ext::FoldSpanExt;
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
-struct FoldSpan {
-    span: Span,
-    kind: FoldKind,
+pub(crate) struct FoldSpan {
+    pub(crate) span: Span,
+    pub(crate) kind: FoldKind,
 }
 
 impl FoldSpan {
     fn sort_key(self) -> (u32, u32, u8) {
         (self.span.start(), self.span.end(), self.kind.sort_key())
-    }
-
-    fn into_lsp(self, line_index: &LineIndex) -> Option<ls_types::FoldingRange> {
-        let range = self.span.to_lsp_range(line_index);
-
-        if range.start.line >= range.end.line {
-            return None;
-        }
-
-        Some(ls_types::FoldingRange {
-            start_line: range.start.line,
-            start_character: Some(range.start.character),
-            end_line: range.end.line,
-            end_character: Some(range.end.character),
-            kind: Some(self.kind.to_lsp_kind()),
-            collapsed_text: None,
-        })
     }
 }
 
@@ -83,7 +64,7 @@ pub fn collect_folding_ranges(
     let line_index = file.line_index(db);
     folds
         .into_iter()
-        .filter_map(|fold| fold.into_lsp(line_index))
+        .filter_map(|fold| fold.to_lsp_folding_range(line_index))
         .collect()
 }
